Precompute XSLT elements after preprocessing
Move the call to xsltStylePreCompute from xsltPrecomputeStylesheet to
later stages of the parsing process. Rename xsltPrecomputeStylesheet
to xsltPreprocessStylesheet. This allows to completely ignore contents
of unknown XSLT elements in forwards-compatible mode.

This is a somewhat dangerous change. If I missed some additional
locations where xsltStylePreCompute must be called, it might result
in errors or performance degradation.

Fixes bug #677901:

https://bugzilla.gnome.org/show_bug.cgi?id=677901