Remove norm:localTime extension function
The length of the input string wasn't checked, resulting in a
minor information leak. This extension function was non-standard
and undocumented and used a custom date format, so it seems best
to remove it.

Note that with the fix to bug #758192, it's possible to convert
between time zones using date:add and date:difference:

    date:add($tz, date:difference($tz, $date-time))

$tz is an arbitrary dateTime in the target time zone. If you
want to convert to the local time zone, set $tz to the current
dateTime returned by date:date-time().

Originally reported to Chromium:

https://crbug.com/670720