Add .travis.yml
For now this is mainly useful if you work on a fork of the libxslt
mirror on GitHub:

    https://github.com/GNOME/libxslt

Start with two build setups:

- GCC with as many GNU extensions disabled as possible, trying to
  emulate a C89 compiler on a POSIX system.

- clang with ASan and UBSan.

The test suite doesn't set an exit code on failure, so log the test
output and grep for unexpected lines. This doesn't work for the Python
tests yet, so they're still disabled.