Fix (then disable) mid-transform space stripping.
Consolidate the mid-transform space stripping to use the same test as the pre-transform space stripping.

But then, space stripping is handled through the `xsltApplyStripSpaces` at both the start of stylesheet application and whenever a document is loaded through the document function.  This means that it doesn't need to be performed mid-transform.

Disable mid-transform space stripping.