libxslt.pc: Do not output private dependencies when linking dynamically
The mathematical library, libm, is a private dependency and should not appear
in command line output unless we have asked the linker to generate
static libraries.

This patch moves the M_LIBS dynamic variable to the Libs.private field in
libxslt.pc and removes $M_LIBS from the XSLT_LIBS variable to prevent the
creation of un-necessary dependencies.