xslt-config: Only output static libraries when '--libs --static' is invoked
xslt-config currently outputs private library dependencies when dynamically
linking (that is, when --libs is invoked). This causes un-necessary dependencies
to be created.

This patch adds a new --static flag to --libs. When both switches are passed to
xslt-config, static library dependencies are passed to the command line. Otherwise,
only the core libxslt libraries are passed with --libs only.