CMake: Relax check for Python 3.x support on Windows
Since the code is now ported to support Python 3.x on Windows, we can enable
building the Python-3.x version of the Python modules on Windows directly.

Not sure about the other platforms that the Python libxslt module supports, so
continue to require unistd.h and have F_GETFL support in fcntl.h on the other
platforms.