CMake: Relax check for enabling crypto support on Windows
The crypto support in libexslt uses the native Windows APIs instead of GCrypt
on Windows, so it is enabled by default in the win32 Makefile build system.

Make it easier for people to enable such support in the CMake builds on
Windows, since we don't need to look for or use GCrypt on Windows when enabling
crypto support.